backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
python-jose==3.5.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
opencv-python==4.12.0.88
mediapipe==0.10.32
python-multipart==0.0.20
